% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-statistics.R
\name{aggr_proj_study_by_pheno}
\alias{aggr_proj_study_by_pheno}
\title{Aggregate project, study info by phenotype column(s)}
\usage{
aggr_proj_study_by_pheno(filter_column, method = 1, con = NULL)
}
\arguments{
\item{filter_column}{one or more filte rcolumns to aggregate by}
}
\description{
Aggregate project, study info by phenotype column(s)
}
\details{
Aggregate project and study information by selected phenotype columns
For example, \code{disease} is phenotype related information that is retained under the
\code{BIOSAMPLE} entity. If we specify \code{filter_column = 'primary_disease'}, then
the function aggregates counts by disease by study. Function also joins in
project id and name
}
\examples{
aggr_proj_study_by_pheno(filter_column = c('sample_cell_type', 'sample_molecule_type'))
          returns
          study_id study_version project_id  project_name          study_name     filter_column     filter_value      total
          528      237             1          1 MMRF CoMMpass MMRF CoMMpass IA13a primary_disease   multiple myeloma  2035
          529      238             1          1 MMRF CoMMpass MMRF CoMMpass IA14a primary_disease   multiple myeloma  2072
}
