% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_get_definitions}
\alias{api_get_definitions}
\title{retrieve custom schema definitions}
\usage{
api_get_definitions(workbook = NULL, con = NULL)
}
\description{
this function retrieves custom schema definitions from scidb (preferred), or from workbook 
(latter method should only be used in development)
}
