% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_register_biosamples}
\alias{api_register_biosamples}
\title{Samples / Biosample}
\usage{
api_register_biosamples(workbook, record, def, entity_to_update = NULL,
  con = NULL)
}
\description{
Samples / Biosample
}
