% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_register_definitions}
\alias{api_register_definitions}
\title{register definitions for dataset}
\usage{
api_register_definitions(df_definitions, record, con = NULL)
}
\description{
register definitions for dataset
}
