% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_register_experimentsets_measurementsets}
\alias{api_register_experimentsets_measurementsets}
\title{Register ExperimentSets and MeasurementSets}
\usage{
api_register_experimentsets_measurementsets(workbook, record, def, con = NULL)
}
\description{
This function differs from api_register_indiv/bios/measurements
that work row by row on Subjects/Samples/Pipelines sheets respectively.
This function needs to find unique set of Experiments from the unique value of `concat` column
in Pipelines sheet, and then uses info in `pipeline_choices` sheet to fill up the 
other necessary information
}
