% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_register_individuals}
\alias{api_register_individuals}
\title{register individuals}
\usage{
api_register_individuals(workbook, record, def, con = NULL)
}
\arguments{
\item{workbook}{workbook object returned by XLConnect:loadWorbook}

\item{record}{record of scidb project_id, dataset_id and dataset_version at which to register individuals}
}
\description{
register individuals
}
