% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_register_measurements}
\alias{api_register_measurements}
\title{Register measurements}
\usage{
api_register_measurements(pipeline_df, bios_df_ref, msmtset_df_ref)
}
\arguments{
\item{pipeline_df}{subselection of Pipeline sheet corresponding to current pipeline}

\item{bios_df_ref}{dataframe containing API entry for biosample data at current \code{dataset_id} (retrieve from reference object)}

\item{msmtset_df_ref}{dataframe containing API entry for current measurementset (retrieve from reference object)}
}
\description{
Excel sheet equivalent of \code{\link{populate_measurements}}
}
