% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_register_ontology_from_definition_sheet}
\alias{api_register_ontology_from_definition_sheet}
\title{automatically register ontology terms}
\usage{
api_register_ontology_from_definition_sheet(workbook = NULL, def = NULL,
  con = NULL)
}
\description{
gather the terms from controlled_vocabulary column, 
categorize by attribute_name, and
assign them ontology_id-s
}
