% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{assign_api_entity}
\alias{assign_api_entity}
\title{Assign entity}
\usage{
assign_api_entity(pipeline_df)
}
\description{
Assign entity based on pipeline and filter choice selection in Pipelines sheet of Excel sheet
}
