% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-synonym.R
\name{build_reference_gene_set}
\alias{build_reference_gene_set}
\title{Build a reference gene set at user specified FeatureSet id}
\usage{
build_reference_gene_set(featureset_id, gene_annotation_file_path = NULL,
  gene_location_file_path = NULL, alias_fields = c("hgnc_id", "symbol",
  "alias_symbol", "prev_symbol", "entrez_id", "ensembl_gene_id", "vega_id",
  "ucsc_id", "ena", "refseq_accession", "ccds_id", "uniprot_ids", "pubmed_id",
  "mgd_id", "rgd_id", "cosmic", "omim_id", "mirbase", "homeodb", "snornabase",
  "bioparadigms_slc", "orphanet", "pseudogene.org", "horde_id", "merops",
  "imgt", "iuphar", "kznf_gene_catalog", "mamit.trnadb", "cd", "lncrnadb",
  "enzyme_id", "intermediate_filament_db"), hierarchy = c("ensembl_gene_id",
  "vega_id", "ucsc_id", "ccds_id", "gene_symbol", "hgnc_id"), con = NULL)
}
\arguments{
\item{featureset_id}{FeatureSet id at which to build a reference gene set}

\item{gene_annotation_file_path}{file providing annotation info (e.g. <hugo/hgnc_complete_set.txt>)}

\item{gene_location_file_path}{gene location info (e.g. <grch38_release85_homosap_gene/newGene.tsv>)}
}
\description{
Build a reference set of genes and synonym at a given featureset_id. Need paths to
<hugo/hgnc_complete_set.txt> for annotation info (param: gene_annotation_file_path)
and <grch38_release85_homosap_gene/newGene.tsv> as gene location info
(param: gene_location_file_path)
}
