% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-statistics.R
\name{calculate_statistics_across_measurementsets}
\alias{calculate_statistics_across_measurementsets}
\title{Calculate summary statistics across measurementsets}
\usage{
calculate_statistics_across_measurementsets(df_measurementset, con = NULL)
}
\description{
Calculate summary statistics across measurementsets
}
\examples{
calculate_statistics_across_measurementsets(
             df_measurementset = search_measurementsets(dataset_id = 31))
calculate_statistics_across_measurementsets(
             df_measurementset = get_measurementsets(measurementset_id = c(2:5)))

}
