% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-statistics.R
\name{calculate_statistics_rnaquantification}
\alias{calculate_statistics_rnaquantification}
\title{Calculate statistics on Gene-Expression data}
\usage{
calculate_statistics_rnaquantification(measurementset_id = NULL,
  decorateResults = TRUE, con = NULL)
}
\description{
Calculate statistics on Gene-Expression data
}
