% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement-statistics.R
\name{calculate_statistics_variant}
\alias{calculate_statistics_variant}
\title{Calculate statistics on Variant data}
\usage{
calculate_statistics_variant(measurementset_id = NULL, con = NULL)
}
\description{
Calculate statistics on Variant data
}
