% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_if_revealgenomics_dataframes_equal}
\alias{check_if_revealgenomics_dataframes_equal}
\title{check if two API results are identical}
\usage{
check_if_revealgenomics_dataframes_equal(df1, df2, entity)
}
\arguments{
\item{df1}{data-frame 1}

\item{df2}{data-frame 2}

\item{entity}{entity for the data-frames that are being compared (e.g. \code{BIOSAMPLE, MEASUREMENTSET} etc.)}
}
\description{
check if two API results are identical
}
\examples{
\dontrun{
b1 = search_biosamples(dataset_id = 1)
b2 = get_biosamples()
b2 = b2[b2$dataset_id ==1, ]
check_if_revealgenomics_dataframes_equal(b1, b2, 'BIOSAMPLE')
}
}
