% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_choices.R
\docType{data}
\name{Choices}
\alias{Choices}
\title{classes to handle different choices sheets in Excel template}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Choices
}
\description{
e.g. pipeline_choices, filter_choices, featureset_choices
multiple matches at a time
}
\keyword{datasets}
