% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{custom_scan}
\alias{custom_scan}
\title{whether to use secure_scan or not}
\usage{
custom_scan()
}
\description{
use this as a switch to choose between
- secure_scan (apply study-level security on dataset_id; must have permissions.dataset_id array)
- scan (just use regular scan of arrays)
}
