% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dao_scidb.R
\name{dao_get_entity}
\alias{dao_get_entity}
\title{dao code for `get_ENTITY`}
\usage{
dao_get_entity(entity, con)
}
\description{
after changes in secure_scan branch, no namespace merging is required (This was the 
faster path used here)
We might get rid of this function and call `get_ENTITY()` directly
}
