% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{dao_search_expression}
\alias{dao_search_expression}
\title{Faster implementation of \code{search_expression} for UI development}
\usage{
dao_search_expression(entity, measurementset, biosample_ref,
  feature = NULL, formExpressionSet = TRUE, con = NULL)
}
\arguments{
\item{biosample_ref}{Reference dataframe containing at least the biosample data for
current study (e.g. by calling
\code{search_biosamples(dataset_id = measurementset$dataset_id)}).
OK if more biosample rows are provided (e.g. by calling
\code{get_biosamples()})}
}
\description{
Here, there is no option to select by specific biosamples
}
