% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{dao_search_rnaquantification}
\alias{dao_search_rnaquantification}
\title{Faster implementation of `search_rnaquantification` for UI development}
\usage{
dao_search_rnaquantification(measurementset, biosample_ref, feature = NULL,
  con)
}
\arguments{
\item{biosample_ref}{Reference dataframe containing at least the biosample data for 
current study (e.g. by calling 
`search_biosamples(dataset_id = measurementset$dataset_id)`). 
OK if more biosample rows are provided (e.g. by calling
`get_biosamples()`)}
}
\description{
Here, there is no option to select by specific biosamples
}
