% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_loader.R
\docType{data}
\name{DataLoader}
\alias{DataLoader}
\title{download features for current featureset}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataLoader
}
\description{
compare with \code{featureset_id} of the current measurementSet
and update the feature stored in \code{reference_object} if there has been a change in \code{featureset_id}
register new features (if any)
}
\details{
return reference_object

For DEBUG only
update feature and feature synonym in reference object
assign feature ids

assume that new features have been registered by this time
assign dataset_id and measurementset_id
}
\keyword{datasets}
