% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_loader.R
\docType{data}
\name{DataLoaderVariantExomic}
\alias{DataLoaderVariantExomic}
\title{assign dataset_id and measurementset_id}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataLoaderVariantExomic
}
\description{
assign dataset_id and measurementset_id
}
\keyword{datasets}
