% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_loader.R
\docType{data}
\name{DataLoaderVariantFormatA}
\alias{DataLoaderVariantFormatA}
\title{loader to try and handle various variant formats
\itemize{
\item currently assumes that features in the data-file are previously registered from a GTF file
\item may use the \code{feature_annotation_df} data.frame to figure out action related to feature matching
}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataLoaderVariantFormatA
}
\description{
loader to try and handle various variant formats
\itemize{
\item currently assumes that features in the data-file are previously registered from a GTF file
\item may use the \code{feature_annotation_df} data.frame to figure out action related to feature matching
}
}
\keyword{datasets}
