% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_loader.R
\docType{data}
\name{DataLoaderVariantGemini}
\alias{DataLoaderVariantGemini}
\title{loader corresponding to GEMINI files
\itemize{
\item May need to use feature-synonym for feature matching, but current implementation is inefficient
\item also has hard-coded links to column names used for feature matching (e.g. \code{gene})
}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataLoaderVariantGemini
}
\description{
loader corresponding to GEMINI files
\itemize{
\item May need to use feature-synonym for feature matching, but current implementation is inefficient
\item also has hard-coded links to column names used for feature matching (e.g. \code{gene})
}
}
\keyword{datasets}
