% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_reader.R
\docType{data}
\name{DataReaderCopyNumberVariantVariableColumns}
\alias{DataReaderCopyNumberVariantVariableColumns}
\title{generic variant loader that aims to work with multiple formats
\itemize{
\item makes sure the mandatory fields for registering variants are named accordingly
\item for biosample matching, preps a column called \code{biosample_name}
\item for feature matching
\itemize{
\item creates a column called \code{scidb_feature_column}
\item stores annotation information in \code{feature_annotation_df} slot
(column name of this data.frame describes whether feature \code{name} or \code{gene_symbol}
should be used for matching)
}
}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataReaderCopyNumberVariantVariableColumns
}
\description{
generic variant loader that aims to work with multiple formats
\itemize{
\item makes sure the mandatory fields for registering variants are named accordingly
\item for biosample matching, preps a column called \code{biosample_name}
\item for feature matching
\itemize{
\item creates a column called \code{scidb_feature_column}
\item stores annotation information in \code{feature_annotation_df} slot
(column name of this data.frame describes whether feature \code{name} or \code{gene_symbol}
should be used for matching)
}
}
}
\keyword{datasets}
