% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaVariantsObject.R
\docType{data}
\name{dnaVariantsObject}
\alias{dnaVariantsObject}
\title{Variant access object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dnaVariantsObject
}
\description{
Variant access object
}
\keyword{datasets}
