% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_unpivot_info_join}
\alias{download_unpivot_info_join}
\title{Download from scidb then unpivot info and join in R}
\usage{
download_unpivot_info_join(qq, arrayname,
  algo_choice_project_primary_id_then_download_attrs_only = FALSE,
  con = NULL)
}
\arguments{
\item{qq}{query on primary array (allied downloaad will be from \code{_INFO} array)}

\item{arrayname}{primary arrayname e.g. \code{BIOSAMPLE, INDIVIDUAL}}

\item{algo_choice_project_primary_id_then_download_attrs_only}{flag to control iquery download mechanism}

\item{con}{connection object (optional when using \code{rg_connect()})}
}
\description{
Download from scidb then unpivot info and join in R
}
\examples{
\dontrun{
download_unpivot_info_join(qq = "filter(secure_scan(gh_secure.BIOSAMPLE), dataset_id = 11)",
                                arrayname = 'BIOSAMPLE')
}
}
