% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{estimate_measurementdata_download_size}
\alias{estimate_measurementdata_download_size}
\title{Estimate downloaded size for measurement data}
\usage{
estimate_measurementdata_download_size(measurementset, units = c("MB",
  "KB", "GB"), con = NULL)
}
\arguments{
\item{measurementset}{measurementset dataframe -- e.g. output of \code{get_measurementset()}
or \code{search_measurementset(dataset_id = ...)}}

\item{units}{return estimated size as either of "\code{KB}", "\code{MB}" (default) or "\code{GB}"}

\item{con}{the connection object (optional)}
}
\description{
(1.) This currently estimates data per pipeline (cannot estimate a subset by feature i.e. genes or biosamples)
(2.) The estimate is a rough estimate e.g. RNA-seq data will have estimate for multiple
columns that are eventually not present in the expression matrix (e.g. \code{dataset_id, dataset_version, 
biosample_id, feature_id, measurementset_id}). Some of these columns are important to link
metadata to the measurement data, and only those are eventually downloaded from the
database.
(3.) For now this is only  used to estimate the data size in the measurement array
(not the feature and any other metadata that might also be downloaded to form a compound
object e.g. \link{\code{search_expression}} returns a Bioconductor ExpressionSet object).
}
