% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionSetObject.R
\docType{data}
\name{expressionSetObject}
\alias{expressionSetObject}
\title{ExpressionSet access object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
expressionSetObject
}
\description{
ExpressionSet access object
}
\keyword{datasets}
