% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_loader.R
\name{feature_matching_level1}
\alias{feature_matching_level1}
\title{Level 1 feature matching}
\usage{
feature_matching_level1(data_df, col_match_ftr_name, fset, feature_df)
}
\description{
Match column in file with feature names
Return result is a structure with matched and unmatched indices
}
