% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_choices.R
\docType{data}
\name{FeaturesetChoices}
\alias{FeaturesetChoices}
\title{extract relevant info pertaining to featureset-choices sheet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FeaturesetChoices
}
\description{
extract relevant info pertaining to featureset-choices sheet
}
\keyword{datasets}
