% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_on_dataset_id_and_version}
\alias{filter_on_dataset_id_and_version}
\title{internal function for `search_METADATA()``}
\usage{
filter_on_dataset_id_and_version(arrayname, dataset_id, dataset_version,
  con = NULL)
}
\description{
internal function for `search_METADATA()``
}
