% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_choices.R
\docType{data}
\name{FilterChoices}
\alias{FilterChoices}
\title{extract relevant info pertaining to filter_choices sheet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FilterChoices
}
\description{
extract relevant info pertaining to filter_choices sheet
}
\keyword{datasets}
