% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-search-index.R
\name{find_dataset_id_by_metadata_value}
\alias{find_dataset_id_by_metadata_value}
\title{Find matching datasets by searching across entities}
\usage{
find_dataset_id_by_metadata_value(metadata_value_df)
}
\arguments{
\item{metadata_value_df}{portion of \code{get_metadata_value()}, must contain the column \code{metadata_value_id}}
}
\description{
Find matching datasets by searching across entities
}
