% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_loader.R
\name{find_matching_featureset}
\alias{find_matching_featureset}
\title{find featureset from Pipeline sheet info}
\usage{
find_matching_featureset(pipeline_df, featureset_link_col, fsets_scidb)
}
\description{
Row or row(s) of Pipeline sheet are used to load data. Unique choice
of featureset name is used to subselect a specific featureset registered
in the database.
}
