% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_project_id_by_grep}
\alias{find_project_id_by_grep}
\title{Returns project_ids by grep on name}
\usage{
find_project_id_by_grep(pattern, con = NULL, ...)
}
\arguments{
\item{pattern}{pattern to search by}

\item{con}{connection object}

\item{...}{additional parameters to grep like \code{ignore.case}, \code{perl}. See documentation for \code{grep}}
}
\description{
Returns project_ids by grep on name
}
