% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-helper-functions.R
\name{formulate_build_literal_query}
\alias{formulate_build_literal_query}
\title{Formulate query to build array from vector}
\usage{
formulate_build_literal_query(vec, value_name = "value_id",
  index_name = "idx")
}
\arguments{
\item{vec}{vector of integers}
}
\description{
Formulate query to build array from vector
}
