% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{formulate_merged_array}
\alias{formulate_merged_array}
\title{formulate merged array query for an entity}
\usage{
formulate_merged_array(entity, con = NULL)
}
\description{
formulate merged array query for an entity
}
\examples{
con = gh_connect2(<ACCESS TO ALL DATA>)
formulate_merged_array_query('RNAQUANTIFICATION', con)
# merge(public.RNAQUANTIFICATION, collaboration.RNAQUANTIFICATION, ....)
con = gh_connect2(<ACCESS TO PUBLIC DATA ONLY>)
formulate_merged_array_query('RNAQUANTIFICATION', con)
# public.RNAQUANTIFICATION
}
