% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{formulate_versioned_selection_query}
\alias{formulate_versioned_selection_query}
\title{Add version filter on top a filter string}
\usage{
formulate_versioned_selection_query(entity, base_selection_query,
  dataset_version)
}
\arguments{
\item{entity}{entity on which filtering is to be run}

\item{base_selection_query}{typically the output of running \code{formulate_base_selection_query}
e.g. (individual_id>=1 AND individual_id<=3)}

\item{dataset_version}{the version at which to filter}
}
\description{
Checks that entity is versioned before adding the version string
}
