% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_entity_from_cache}
\alias{get_entity_from_cache}
\title{get_ENTITY for cached entities}
\usage{
get_entity_from_cache(entitynm, id, updateCache, con = NULL)
}
\arguments{
\item{entitynm}{entity name}

\item{id}{the id-s to be selected (can be \code{NULL})}

\item{cache_df}{typically result of a \code{get_entity_from_cache()} call}
}
\description{
Currently works for ONTOLOGY, DEFINITION, VARIANT_KEY
}
