% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schema.R
\name{get_entity_from_entity_id}
\alias{get_entity_from_entity_id}
\title{Get entity name}
\usage{
get_entity_from_entity_id(entity_id)
}
\description{
Get entity name from entity id
}
\examples{
get_entity_from_entity_id(1002)            # 'PROJECT'
get_entity_from_entity_id(8002)            # 'VARIANT'
get_entity_from_entity_id(c(1002, 8002))   # c('PROJECT', 'VARIANT')
\dontrun{
get_entity_from_entity_id(c(1002, -1))     # Error
}
}
