% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{get_experiment}
\alias{get_experiment}
\title{retrieve all the experiments available to logged in user}
\usage{
get_experiment()
}
\description{
joins Measurement, Dataset (for study category field), and ExperimentSet arrays
to return experiment information to user
}
