% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{get_experiments}
\alias{get_experiments}
\title{retrieve all the experiments available to logged in user}
\usage{
get_experiments(con = NULL)
}
\description{
joins `Measurement`, `Dataset` (for `study category` field if available), and `ExperimentSet` arrays
to return experiment information to user
}
\examples{
experiments = get_experiments()
cat("Categorization of experiments by major type\\n")
table(experiments$measurement_entity)
cat("Categorization of experiments by sub type\\n")
table(paste(experiments$measurement_entity, experiments$name, sep = ": "))
cat("Categorization of experiments by sub type\\n")
table(experiments$`study category`)

}
