% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_genelist_gene_symbols}
\alias{get_genelist_gene_symbols}
\title{Gene list gene symbols visible to user}
\usage{
get_genelist_gene_symbols(con = NULL)
}
\value{
data.frame containing \code{genelist_id} and \code{gene_symbol}
}
\description{
Get data.frame containing summary of all gene-symbols (and genelist_id-s)
from genelist-s visible to user
}
