% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schema.R
\name{get_mandatory_fields_for_register_entity}
\alias{get_mandatory_fields_for_register_entity}
\title{mandatory fields (internal function)}
\usage{
get_mandatory_fields_for_register_entity(arrayname)
}
\description{
return mandatory fields while registering an entity
}
\details{
'DATASET': attributes
other secure metadata entities (e.g. 'INDIVIDUAL', 'BIOSAMPLE', 'BIOSAMPLE'): 'dataset_id' and attributes
public metadata entities (e.g. 'ONTOLOGY'): attributes
feature entities: attributes
}
