% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{grant_admin_access}
\alias{grant_admin_access}
\title{grant scidbadmin access to studies 1:N}
\usage{
grant_admin_access(con, max_study_id)
}
\description{
workaround till https://github.com/Paradigm4/secure_scan/issues/7 is fixed
}
