% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{grant_initial_access}
\alias{grant_initial_access}
\title{grant initial access to a user}
\usage{
grant_initial_access(con = NULL, user_name)
}
\description{
function to be called while onboarding a user
can only be called by scidbadmin
}
