% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_object_in_scidb.R
\name{init_r_object_cache}
\alias{init_r_object_cache}
\title{Initialize R object cache array}
\usage{
init_r_object_cache(con = NULL)
}
\arguments{
\item{con}{connection object}
}
\description{
Initialize R object cache array
}
\examples{
\dontrun{
  con_root = rg_connect2(username = 'scidbadmin') # has to be scidbadmin
  revealgenomics:::init_r_object_cache(con = con_root)
}
}
