% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schema.R
\name{is_entity_cached}
\alias{is_entity_cached}
\title{is entity cached}
\usage{
is_entity_cached(entitynm)
}
\description{
check if entity is cached
}
\details{
if no value is provided in schema file, then the entity is potentially not cached
}
