% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{jnjscidb}
\alias{jnjscidb}
\alias{jnjscidb-package}
\title{JNJ SciDB Toolbox for R}
\description{
Available package options (you can override the defaults with the \code{options} function):
}
\details{
\itemize{
\item \code{jnjscidb.dummy_option=7999}
}
Placeholder for documentation
}
\seealso{
\code{\link{get_individuals}}, \code{\link{search_individuals}},
}
