% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-helper.R
\name{list_ncol_in_aux_column}
\alias{list_ncol_in_aux_column}
\title{List count of sub-columns per row in a column that is a JSON formatted string}
\usage{
list_ncol_in_aux_column(df1, aux_info_colname)
}
\arguments{
\item{df1}{A dataframe}

\item{aux_info_colname}{Column name of JSON formatted string}
}
\description{
List count of sub-columns per row in a column that is a JSON formatted string
}
\examples{
df1 = data.frame(name = c("Alice", "Bob"),
                 details__JSON = c('[{  "boolean": true,   "null": null,   "number": 123,   "string":  "Hello World" }]',
                                   '[{  "boolean": false,                  "number": 456,   "differentString":  "Hi World" }]'), 
                 stringsAsFactors = FALSE)
list_ncol_in_aux_column(df1, 'details__JSON')
}
