% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-helper.R
\name{list_nrow_in_aux_column}
\alias{list_nrow_in_aux_column}
\title{List count of sub-rows per row in a column that is a JSON formatted string}
\usage{
list_nrow_in_aux_column(df1, aux_info_colname)
}
\arguments{
\item{df1}{A dataframe}

\item{aux_info_colname}{Column name of JSON formatted string}
}
\description{
List count of sub-rows per row in a column that is a JSON formatted string
}
\examples{
df1 = data.frame(name = c("Alice", "Bob"),
                 visits__JSON = c('[{"visit_num":1,"condition":"Baseline"}] ',
                                   '[{"visit_num":1,"condition":"Baseline"},{"visit_num":2,"condition":"Relapse"}] '), 
                 stringsAsFactors = FALSE)
list_nrow_in_aux_column(df1, 'visits__JSON')
}
