% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-helper.R
\name{load_helper_column_rename}
\alias{load_helper_column_rename}
\title{helper function to rename columns}
\usage{
load_helper_column_rename(dfx, scidb4gh_fields, worksheet_fields)
}
\description{
rename columns from excel template into scidb4gh format
}
