% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-helper.R
\name{load_helper_do_entity_specific_work}
\alias{load_helper_do_entity_specific_work}
\title{Record all entity specific work per entity in one place}
\usage{
load_helper_do_entity_specific_work(data_df, entity, record, con = NULL)
}
\description{
Record all entity specific work per entity in one place
}
