% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-helper.R
\name{load_helper_handle_duplicates}
\alias{load_helper_handle_duplicates}
\title{Drop duplicate rows before importing}
\usage{
load_helper_handle_duplicates(data_df, entity)
}
\description{
Drop duplicate rows before importing
}
