% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-helper.R
\name{load_helper_prepare_dataframe}
\alias{load_helper_prepare_dataframe}
\title{Wrapper function combining common actions for preparing dataframe for loading to scidb}
\usage{
load_helper_prepare_dataframe(workbook, record, def, sheetName, entityName,
  worksheet_fields, con = NULL)
}
\description{
Applicable to \code{Individuals}, \code{Biosamples} and \code{Measurements} that are directly loaded
from each row of \code{Subjects}, \code{Samples} and \code{Pipelines} sheet
}
