% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-helper.R
\name{load_helper_replace_local_ids}
\alias{load_helper_replace_local_ids}
\title{get rid of local indexes}
\usage{
load_helper_replace_local_ids(data_df, record)
}
\description{
NOTE: need to run this function at the right point, where do not need to access
local id-s anymore
}
