% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schema.R
\name{mandatory_fields}
\alias{mandatory_fields}
\title{mandatory fields}
\usage{
mandatory_fields()
}
\description{
return mandatory fields while registering entities
}
\details{
'DATASET': attributes
other secure metadata entities (e.g. 'INDIVIDUAL', 'BIOSAMPLE', 'BIOSAMPLE'): 'dataset_id' and attributes
public metadata entities (e.g. 'ONTOLOGY'): attributes
feature entities: attributes
}
