% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dataframe_loader.R
\name{match_features}
\alias{match_features}
\title{convenience function for matching features between database and file}
\usage{
match_features(features_in_file, df_features_db, feature_type, column_in_db)
}
\description{
convenience function for matching features between database and file
}
