% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-helper.R
\name{merge_aux_info_from_df}
\alias{merge_aux_info_from_df}
\title{Get auxiliary information from a dataframe, and merge as JSON strings into master dataframe}
\usage{
merge_aux_info_from_df(master_df, aux_df, aux_info_colname, by.x, by.y,
  verbose_merge = FALSE)
}
\arguments{
\item{master_df}{Dataframe containing master merge information (e.g. names of individuals)}

\item{aux_df}{Dataframe containing auxiliary information}

\item{aux_info_colname}{Column name in which to store auxiliary information as JSON objects}

\item{by.x}{Merge key on master dataframe}

\item{by.y}{Merge key on auxiliary dataframe}

\item{verbose_merge}{boolean flag to print output while merging}
}
\description{
Get auxiliary information from a dataframe, and merge as JSON strings into master dataframe
}
\examples{
individual_df = data.frame(name = c('Alex', 'Bob'), age = c(44, 33), stringsAsFactors = FALSE)
visit_df = data.frame(individual = c('Alex', 'Alex', 'Bob'),
                    visit_num = c(1, 2, 1), 
                    reason = c('Baseline', 'Relapse', 'Baseline'),
                    temperature = c(99.5, 97.9, 99.2))
merge_aux_info_from_df(master_df = individual_df, 
                       aux_df = visit_df, 
                       aux_info_colname = 'visits__JSON', 
                       by.x = 'name', by.y = 'individual')
}
