% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-helper.R
\name{merge_aux_info_from_file}
\alias{merge_aux_info_from_file}
\title{Get auxiliary information from a file, and merge as JSON strings into master dataframe}
\usage{
merge_aux_info_from_file(master_df, aux_file_path, aux_info_colname,
  by.master, by.aux, verbose_read = TRUE, verbose_merge = FALSE)
}
\arguments{
\item{master_df}{Dataframe containing master merge information (e.g. names of individuals)}

\item{aux_file_path}{File from which to read auxiliary information}

\item{aux_info_colname}{Column name in which to store auxiliary information as JSON objects}

\item{verbose_read}{boolean flag to print output on reading file}

\item{verbose_merge}{boolean flag to print output while merging}

\item{by.x}{Merge key on master dataframe}

\item{by.y}{Merge key on auxiliary dataframe}
}
\description{
This is a wrapper for merge_aux_info_from_df() in that it fails gracefully if there is an
error at any stage. Also it prints some output as it goes along.
}
