% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{myExcelLoader}
\alias{myExcelLoader}
\title{Load Excel workbook into memory}
\usage{
myExcelLoader(filename, use_readxl = TRUE, silently = FALSE)
}
\arguments{
\item{filename}{path to Excel workbook}

\item{use_readxl}{use readxl package (default), otherwise XLConnect package}

\item{silently}{spew log messages while loading workbook (logging turned on by default)}
}
\description{
To be used in conjunction with \code{myExcelReader()}
`
}
