% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{myExcelReader}
\alias{myExcelReader}
\title{Read sheet from Excel workbook}
\usage{
myExcelReader(workbook, sheet_name)
}
\arguments{
\item{workbook}{Workbook must be of type `XLConnect::workbook` (loaded by `XLConnect::loadWorkbook()`) 
or of type `list` (loaded by `scidb4gh::myExcelLoader()`)}

\item{sheet_name}{name of sheet to read from workbook}
}
\description{
To be used in conjunction with `myExcelLoader()`.
}
\details{
Note usage of the useCachedValues parameter
otherwise, was throwing warnings about the concatenation field
}
