% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{names_to_numbered_vec_by_uniqueness}
\alias{names_to_numbered_vec_by_uniqueness}
\title{names to list of numbers by uniqueness}
\usage{
names_to_numbered_vec_by_uniqueness(names_vec)
}
\description{
function to convert a list of names to
a numbered vector \code{1:N} where \code{N} is the number of unique names
(each unique name has a different number)
}
\examples{
names_to_numbered_vec_by_uniqueness(c('a', 'b', 'a', 'c')) # returns: 1, 2, 1, 3
names_to_numbered_vec_by_uniqueness(c(11, 35, 44, 11, 35, 66)) # returns: 1, 2, 3, 1, 2, 4
}
