% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_choices.R
\docType{data}
\name{PipelineChoices}
\alias{PipelineChoices}
\title{extract relevant info pertaining to pipeline_choices sheet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PipelineChoices
}
\description{
extract relevant info pertaining to pipeline_choices sheet
}
\keyword{datasets}
