% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-register-measurementdata.R
\name{register_copynumbervariant_variable_columns}
\alias{register_copynumbervariant_variable_columns}
\title{Register CNV data of variable columns subtype}
\usage{
register_copynumbervariant_variable_columns(df1, measurementset,
  only_test = FALSE, con = NULL)
}
\description{
Function to register CNV data when the underlying data is a table of variable columns
}
