% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_excel_loader.R
\name{register_entities_workbook}
\alias{register_entities_workbook}
\title{register entities via list of R data-frames}
\usage{
register_entities_workbook(workbook, register_upto_entity = c("all",
  "ONTOLOGY", "DEFINITION", "INDIVIDUAL", "BIOSAMPLE", "MEASUREMENTSET",
  "MEASUREMENT"), entity_to_update = c(NULL, "PROJECT", "DATASET",
  "INDIVIDUAL", "BIOSAMPLE", "EXPERIMENTSET", "MEASUREMENTSET", "all"),
  register_measurement_entity = c("all", "RNAQUANTIFICATION", "VARIANT",
  "FUSION", "PROTEOMICS", "COPYNUMBER_SEG", "COPYNUMBER_MAT"),
  pipeline_name_filter = NULL, con = NULL)
}
\arguments{
\item{register_upto_entity}{use this parameter to restrict load up to an user-specified entity.
e.g. you can load up to the following and not load the rest (1) \code{ONTOLOGY}, or
(2) \code{DEFINITION} (this includes \code{PROJECT} and \code{DATASET}), or
(3) \code{INDIVIDUAL}, and so on. The ordering is same
as in the parameter definition below.
Default value is \code{all} i.e. all entities are loaded}

\item{entity_to_update}{enable excel sheet loader to update one entity at a time
- applies to project, dataset, individual, biosample, experimentset,
measurementset. If \code{all} is supplied, then all metadata entities
will be updated}

\item{register_measurement_entity}{use this parameter to restrict which measurement-data types
should be loaded by the function.
Default value is \code{all} i.e. all measurement-types are loaded}

\item{pipeline_name_filter}{use this filter to restrict data-loading to pipelines with fully/partially
matched names. Default value is \code{NULL} (i.e. no checking by name done)}
}
\description{
Register entities using list of R data-frames
list of R data-frames has complete data about metadata (\code{PROJECT}, \code{DATASET}, \code{INDIVIDUAL}, ...) ,
and links to files storing the \code{MeasurementData} entities
(e.g. \code{RNASeq}, \code{Variant} etc.)
}
