% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register_entity_flex_fields}
\alias{register_entity_flex_fields}
\title{Register info for entities into entity wide array}
\usage{
register_entity_flex_fields(df1, idname, arrayname, con = NULL)
}
\description{
Curate attributes (column names) into one array, values (column values) in another,
and link all into one common entity-side array
}
