% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-helper-filetypes.R
\name{register_expression_file_cufflinks__DEPRECATED}
\alias{register_expression_file_cufflinks__DEPRECATED}
\title{DEPRECATED}
\usage{
register_expression_file_cufflinks__DEPRECATED(filepath, separator, dataset_id,
  dataset_version, biosample_name, featureset_id, rnaquantificationset_id,
  biosample_df = NULL, feature_df = NULL)
}
\arguments{
\item{biosample_name}{unique sample name at which to register biosample}

\item{featureset_id}{featureset ID in which to look for features}

\item{biosample_df}{(optional) dataframe consisting biosamples of current study}

\item{feature_df}{(optional) dataframe consisting of features on current featureset_id}
}
\description{
stop("Code integrated into: DataReaderRNASeqCufflinks::load_data_from_file(). Use that function instead")
  
  Eventually this one function should make `register_expression_file_cufflinks_gene` redundant
}
