% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-register-measurementdata.R
\name{register_expression_matrix}
\alias{register_expression_matrix}
\title{Upload expression matrix file}
\usage{
register_expression_matrix(filepath, measurementset_id, featureset_id,
  feature_type, dataset_version = NULL, only_test = FALSE, con = NULL)
}
\description{
filepath must exist on SciDB server. Altrnatively use
\code{register_expression_matrix_client}
}
