% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-measurement-upload.R
\name{register_expression_matrix_client}
\alias{register_expression_matrix_client}
\title{Upload expression matrix file}
\usage{
register_expression_matrix_client(filepath, measurementset, featureset,
  biosample_ref = NULL, feature_ref = NULL, con = NULL)
}
\description{
filepath can exist on SciDB server or client
`register_expression_matrix_client`
}
