% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-gtf.R
\name{register_features_from_gtf_file}
\alias{register_features_from_gtf_file}
\title{Register features from GTF file}
\usage{
register_features_from_gtf_file(featureset_name, gtf_filepath, nSkip,
  register_transcripts = c(FALSE, TRUE), con = NULL)
}
\arguments{
\item{featureset_name}{Exact featureset name at which to register features}

\item{gtf_filepath}{Path to GTF file}

\item{nSkip}{Number of rows to skip from top of GTF file}

\item{register_transcripts}{whether to register_transcripts (TRUE) or not (FALSE)}
}
\value{
list containing feature id-s of registered genes and transcripts (if selected)
}
\description{
Register features from GTF file
}
