% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-register-cached.R
\name{register_gene_symbol}
\alias{register_gene_symbol}
\title{register gene_symbols}
\usage{
register_gene_symbol(df, only_test = FALSE, con = NULL)
}
\arguments{
\item{df}{data-frame containing gene_symbol and full name}
}
\value{
gene_symbol_id assigned by the data-base
}
\description{
register gene_symbols
}
