% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register_genelist}
\alias{register_genelist}
\title{Register genelist}
\usage{
register_genelist(genelist_name = NULL, genelist_description = NULL,
  isPublic = TRUE, df = NULL, only_test = FALSE, con = NULL)
}
\arguments{
\item{genelist_name}{the name of the geneliest}

\item{genelist_description}{the description for the genelist}

\item{isPublic}{bool to denote if the genelist is public}

\item{df}{(optional) a data-frame containing [name, description] of the genelist}
}
\description{
Preferred method of registering genelist-s is to 
provide one genelist_name and genelist_description
at a time.
}
\details{
Alternatively, you can supply a data.frame with [name, description] 
of one or more genelist(s)
}
