% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register_genelist_gene}
\alias{register_genelist_gene}
\title{Register gene symbols in a gene-list}
\usage{
register_genelist_gene(genelist_id = NULL, gene_symbols = NULL, df = NULL,
  only_test = FALSE, con = NULL)
}
\arguments{
\item{genelist_id}{the id of the genelist (returned by \code{register_genelist()})}

\item{gene_symbols}{the gene-symbols to be stored in a gene-list (e.g. \code{c('TSPAN6', 'KCNIP2', 'CFAP58', 'GOT1', 'CPN1', 'PSIP1P1')})}

\item{df}{(optional) a data-frame containing \link{genelist_id, gene_symbol}}
}
\description{
Preferred method of registering gene symbols in a genelist is to
provide the target genelist_id and the symbols to be registered
(one genelist at a time)
}
\details{
Alternatively, one can supply a data.frame with \link{genelist_id, gene_symbol}
}
\examples{
register_genelist_gene(genelist_id = 11, # must already exist in `genelist` table
                       gene_symbols = c('TSPAN6', 'KCNIP2'))
                       
}
