% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-register-measurementdata.R
\name{register_in_memory_matrix}
\alias{register_in_memory_matrix}
\title{Register in-memory matrix}
\usage{
register_in_memory_matrix(mat, measurementset)
}
\arguments{
\item{mat}{in-memory matrix of feature_id (rows) by biosample_id (columns) and (expression/Proteomics/...) value}

\item{measurementset}{data.frame containing information about MeasurementSet}
}
\description{
Register in-memory matrix (Expression, Proteomics etc.) into corresponding entity
}
