% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register_mandatory_and_flex_fields}
\alias{register_mandatory_and_flex_fields}
\title{Register the mandatory fields and update lookup}
\usage{
register_mandatory_and_flex_fields(df, arrayname, con = NULL)
}
\description{
Wrapper function that
(1) registers mandatory fields,
(2) registers flex fields
}
