% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{register_measurements}
\alias{register_measurements}
\title{register all measurements for a specific dataset}
\usage{
register_measurements(dataset_id, dataset_version, con = NULL)
}
\description{
Work in progress
Once finished, this function should eventually be used instead of the script <load-experiment.R>
The function would be called after loading the data for each dataset
It compiles information for all measerement entities, and inserts the relevant
information into the MEASUREMENT entity (at this point, this is called the EXPERIMENT entity,
but the name should be changed. A MEASUREMENT combines both the Experiment and the 
Pipeline information)
}
