% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_object_in_scidb.R
\name{register_r_object_at_key}
\alias{register_r_object_at_key}
\title{Upload an R object at key}
\usage{
register_r_object_at_key(r_obj, key, con = NULL)
}
\arguments{
\item{r_obj}{R object to be stored in scidb binary form}

\item{key}{key to associate with the R object}

\item{con}{connection object}
}
\description{
Store an R object in scidb and associate it with user provided key
}
\examples{
\dontrun{
  rg_connect(username = 'scidbadmin') # or any user with access to gh_secure namespace
  revealgenomics:::register_r_object_at_key(r_obj = c(1:10), key = "asdf")
  # To retrieve data, run
  revealgenomics:::retrieve_r_object_at_key(key = "asdf")
}
}
