% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register_tuple_update_lookup}
\alias{register_tuple_update_lookup}
\title{Register the mandatory fields and update lookup}
\usage{
register_tuple_update_lookup(df, arrayname, updateLookup, con = NULL)
}
\description{
Wrapper function that 
(1) registers mandatory fields, 
(2) updates lookup array (based on flag), and 
(3) registers flex fields
}
