% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-helper-functions.R
\name{remove_old_versions_for_entity}
\alias{remove_old_versions_for_entity}
\title{remove old array versions associated with an entity}
\usage{
remove_old_versions_for_entity(entitynm, con = NULL)
}
\description{
remove old array versions associated with an entity
}
