% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-helper-functions.R
\name{remove_versions}
\alias{remove_versions}
\title{remove old versions associated with an array}
\usage{
remove_versions(arayname, con = NULL)
}
\description{
Retain the last N_THRESH versions
}
