% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_object_in_scidb.R
\name{retrieve_r_object_at_key}
\alias{retrieve_r_object_at_key}
\title{Retrieve R object associated with key}
\usage{
retrieve_r_object_at_key(key, con = NULL)
}
\arguments{
\item{key}{key associated with the R object}

\item{con}{connection object}
}
\description{
Retrieve R object associated with key
}
\examples{
\dontrun{
  rg_connect(username = 'scidbadmin') # or any user with access to gh_r_object_store namespace
  revealgenomics:::retrieve_r_object_at_key(key = "asdf")
  revealgenomics:::retrieve_r_object_at_key(key = "KeyDoesNotExist")
  # No R object found at key: KeyDoesNotExist
}
}
