% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-register-measurementdata.R
\name{return_sub_indices}
\alias{return_sub_indices}
\title{Return sub indices}
\usage{
return_sub_indices(bigN, len_subelem)
}
\arguments{
\item{bigN}{\code{c(1: bigN)} is the vector that needs to be split}

\item{len_subelem}{The maxiumum length of one of the split vectors}
}
\description{
Split a vector \code{1:bigN} into list of vectors each of length < \code{len_subelem}
}
\examples{
return_sub_indices(bigN = 10, len_subelem = 4)
## [[1]]
## [1] 1 2 3 4
## 
## [[2]]
## [1] 5 6 7 8
## 
## [[3]]
## [1]  9 10
}
