% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{revealgenomics}
\alias{revealgenomics}
\alias{revealgenomics-package}
\title{GA4GH inspired and SciDB backed R API for genomic data}
\description{
Available package options (you can override the defaults with the \code{options} function):
}
\details{
\itemize{
\item \code{revealgenomics.dummy_option=7999}
}
Placeholder for documentation
}
\seealso{
\code{\link{get_individuals}}, \code{\link{search_individuals}},
}
