% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{run_quality_control_checks}
\alias{run_quality_control_checks}
\title{Run QC checks on data loaded via Excel file}
\usage{
run_quality_control_checks(measurementset_id = NULL)
}
\arguments{
\item{measurementset_id}{one / set of \code{measurementset_id}-s to restrict QC checking to.
If not supplied, then QC checks will be carried out on all
\code{measurementset_id}-s that are available in MeasurementSet entity}
}
\description{
Run quality control checks on data loaded via Excel file. Make sure Measurement array is linked
properly to (1) corresponding entries in Biosample entity, (2) corresponding entries
in MeasurementImportFile entity
}
