% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_copy_number_variant}
\alias{search_copy_number_variant}
\title{Search copy number variant data}
\usage{
search_copy_number_variant(measurementset, feature = NULL,
  biosample = NULL, con = NULL)
}
\arguments{
\item{measurementset}{(Mandatory) dataframe containing pipeline information;
typically output of a
\code{get_measurementsets(measurementset_id = ...)} or
\code{search_measurementsets(dataset_id = ..)} call}

\item{feature}{(Optional) dataframe containing feature information;
typically output of a
\code{search_features(gene_symbol = ...)} call.
If not specified, function returns all features available
by other search parameters}

\item{biosample}{(Optional) dataframe containing biosample information;
typically output of a
\code{search_biosamples(dataset_id = ..)} call.
If not specified, function returns all biosamples available
by other search parameters}

\item{con}{(Optional) database connection object; typically output of \code{rg_connect2()}
call. If not specified, connection object is formulated from internally stored
values of \code{rg_connect()} call}
}
\description{
Function to search copy number variant data
}
