% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_copynumbervariant_variable_columns}
\alias{search_copynumbervariant_variable_columns}
\title{Search copy number variant subtype (variable columns)}
\usage{
search_copynumbervariant_variable_columns(measurementset,
  biosample = NULL, feature = NULL, dataset_lookup_ref = NULL,
  con = NULL)
}
\description{
Search copy number variant subtype (variable columns)
}
