% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_expression_by_one_measurementset_zero_or_more_features}
\alias{search_expression_by_one_measurementset_zero_or_more_features}
\title{Subpath for \code{search_expression}}
\usage{
search_expression_by_one_measurementset_zero_or_more_features(entity,
  measurementset, biosample_ref, feature = NULL,
  formExpressionSet = TRUE, con = NULL)
}
\arguments{
\item{biosample_ref}{Reference dataframe containing at least the biosample data for
current study (e.g. by calling
\code{search_biosamples(dataset_id = measurementset$dataset_id)}).
OK if more biosample rows are provided (e.g. by calling
\code{get_biosamples()})}
}
\description{
Faster implementation of \code{search_expression} for one measurementset (Pipeline), and multiple features. Here, there is no option to select by specific biosamples
}
