% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_feature_by_synonym}
\alias{search_feature_by_synonym}
\title{Search features by synonym}
\usage{
search_feature_by_synonym(synonym, id_type = NULL, featureset_id = NULL,
  updateCache = TRUE)
}
\arguments{
\item{synonym:}{A name for a gene by any convention e.g. ensembl_gene_id, entrez_id, vega_id}

\item{id_type:}{(Optional) The id type by which to search e.g. ensembl_gene_id, entrez_id, vega_id}

\item{featureset_id:}{(Optional) The featureset within which to search}
}
\value{
feature(s) associated with provided synonym
}
\description{
Search features by synonym
}
