% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_features}
\alias{search_features}
\title{Search features}
\usage{
search_features(gene_symbol = NULL, feature_type = NULL,
  featureset_id = NULL, con = NULL)
}
\arguments{
\item{gene_symbol}{(optional) which gene symbol(s) e.g. \code{c('EGFR', 'KRAS')} / primary symbol(s) e.g. \code{c('TIME', 'CELL.LENGTH')} to search by}

\item{feature_type}{(optional) subselect by feature_type. Ranges between \code{c('gene', 'probeset', 'transcript', 'protein_probe', 'cytof')}}
}
\description{
Search features
}
