% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_fusion}
\alias{search_fusion}
\title{Search Fusion data}
\usage{
search_fusion(measurementset, biosample = NULL, feature = NULL,
  dataset_lookup_ref = NULL, con = NULL)
}
\description{
Search Fusion data
}
