% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_individuals}
\alias{search_individuals}
\title{Search individuals by dataset}
\usage{
search_individuals(dataset_id = NULL, requested_attributes = NULL,
  dataset_version = NULL, all_versions = FALSE, con = NULL)
}
\description{
\code{search_individuals()} can be used to retrive
all individuals in a particular dataset
}
