% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_metadata_attrkey}
\alias{search_metadata_attrkey}
\title{Search metadata attributes by entity id}
\usage{
search_metadata_attrkey(entity_id, updateCache = FALSE, con = NULL)
}
\arguments{
\item{entity_id}{id of API entity as assigned in \code{SCHEMA.yaml} file. You can list all
\code{entity_id}-s by running the function \code{get_entity_info()}}
}
\description{
Search metadata attributes by entity id
}
