% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_metadata_value}
\alias{search_metadata_value}
\title{Search metadata value array by values}
\usage{
search_metadata_value(metadata_value = NULL,
  ontology_category = "uncategorized", updateCache = FALSE,
  con = NULL)
}
\arguments{
\item{metadata_value}{values to search by. If \code{NULL}, return all values at that category}

\item{ontology_category}{ontology category to search by (default: "uncategorized"). If \code{NULL}, search values across all categories}
}
\description{
Search metadata value array by values
}
