% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_ontology_category}
\alias{search_ontology_category}
\title{Search ontology category by categories}
\usage{
search_ontology_category(ontology_category, updateCache = FALSE,
  con = NULL)
}
\arguments{
\item{ontology_category}{one or more categories to search by}
}
\description{
Search by categories and return \code{data.frame} of \code{ontology_category_id, ontology_category}
}
