% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_rnaquantification}
\alias{search_rnaquantification}
\title{Deprecated function to search gene expression data}
\usage{
search_rnaquantification(...)
}
\arguments{
\item{...}{refer parameters for function \code{\link{search_expression}}}
}
\description{
This function is kept around for backward compatibility.
Use \code{\link{search_expression}} instead.
}
\details{
Function to search gene expression data array; allows slicing across multiple
dimensions. However \code{measurementset} (i.e. pipeline) must be supplied
(currently allows searching one pipeline at a time)
}
