% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_variant}
\alias{search_variant}
\title{search variants}
\usage{
search_variant(measurementset, biosample = NULL, feature = NULL,
  autoconvert_characters = TRUE, con = NULL)
}
\arguments{
\item{measurementset}{pipeline dataframe (mandatory)}
}
\description{
Search variants by pipeline, sample and/or  feature
}
