% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_variants}
\alias{search_variants}
\title{Deprecated function for searching variant data}
\usage{
search_variants(...)
}
\arguments{
\item{...}{refer parameters for function \code{\link{search_variant}}}
}
\description{
This function is kept around for backward compatibility.
Use \code{\link{search_variant}} instead.
}
\details{
Function to search variant data array; allows slicing across multiple
dimensions. \code{measurementset} (i.e. pipeline) must be supplied
(currently allows searching one pipeline at a time)
}
