% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_variants_scidb}
\alias{search_variants_scidb}
\title{Inner function for searching variants}
\usage{
search_variants_scidb(arrayname, measurementset_id, biosample_id = NULL,
  feature_id = NULL, dataset_version, con = NULL)
}
\description{
Inner function for searching variants
}
