% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_versioned_secure_metadata_entity}
\alias{search_versioned_secure_metadata_entity}
\title{internal function for search_METADATA()}
\usage{
search_versioned_secure_metadata_entity(entity, dataset_id,
  dataset_version, all_versions, con = NULL)
}
\description{
internal function for \code{search_individuals()}, \code{search_biosamples()} etc.
search of a metadata entry by \code{dataset_id}
}
