% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_versioned_secure_metadata_entity_by_requested_attributes}
\alias{search_versioned_secure_metadata_entity_by_requested_attributes}
\title{internal function for \code{search_METADATA()} by set of requested attributes}
\usage{
search_versioned_secure_metadata_entity_by_requested_attributes(entity,
  requested_attributes, dataset_id, dataset_version, all_versions, con)
}
\description{
internal function for \code{search_individuals()}, \code{search_biosamples()} etc.
search of a metadata entry by \code{requested_attributes}
}
