% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_from_1d_entity}
\alias{select_from_1d_entity}
\title{internal function for get_METADATA()}
\usage{
select_from_1d_entity(entitynm, id, dataset_version = NULL,
  mandatory_fields_only = FALSE, con = NULL)
}
\description{
function to select from metadata entities. Used by 
`get_project()` (does not have dataset_version as a dimension), and
`get_dataset()`, `get_biosamples()` etc. (that have dataset_version as a dimension)
}
