% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{set_permissions}
\alias{set_permissions}
\title{set per study access permissions for a user}
\usage{
set_permissions(con = NULL, user_name, dataset_names, allowed)
}
\description{
to be run only by scidbadmin, or user 
with Read / Write capability to permissions and secured namespaces
}
\details{
can supply multiple study names at a time
}
