% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{show_dataset_permissions}
\alias{show_dataset_permissions}
\title{show study permissions}
\usage{
show_dataset_permissions(con = NULL, dataset_name)
}
\description{
to be run only by scidbadmin, or user 
with Read capability to permissions and secured namespaces
}
