% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{show_roles_for_user}
\alias{show_roles_for_user}
\title{Function to show rules for user}
\usage{
show_roles_for_user(con)
}
\description{
Since \code{iquery(con$db, "show_roles_for_user(...)", return=TRUE)} does not work
Refer https://github.com/Paradigm4/accelerated_io_tools/issues/33
}
