% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-api.R
\name{show_user_permissions}
\alias{show_user_permissions}
\title{show user permissions}
\usage{
show_user_permissions(con = NULL, user_name)
}
\description{
to be run only by scidbadmin, or user 
with Read capability to permissions and secured namespaces
}
