% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_assign_experiment_entity}
\alias{template_helper_assign_experiment_entity}
\title{Assign experiment entity}
\usage{
template_helper_assign_experiment_entity(experiment_name)
}
\arguments{
\item{experiment_name}{from \code{data_subtype} column from Excel template}
}
\description{
Assign experiment entity
}
