% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_assign_measurement_entity}
\alias{template_helper_assign_measurement_entity}
\title{Assign measurement entity}
\usage{
template_helper_assign_measurement_entity(pipeline_df)
}
\description{
Assign measurement entity based on pipeline and filter choice selection in Pipelines sheet of Excel sheet
}
