% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_convert_names}
\alias{template_helper_convert_names}
\title{Convert external (human-readable names) to API internal names}
\usage{
template_helper_convert_names(api_name = NULL, external_name = NULL)
}
\description{
external                   API
Gene Expression            RNAQUANTIFICATION
Variant                    VARIANT
Rearrangement              FUSION
}
