% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_enforce_columns_defined}
\alias{template_helper_enforce_columns_defined}
\title{Enforce that columns in any data sheet have definitions in Definitions sheet}
\usage{
template_helper_enforce_columns_defined(data_df, definitions)
}
\description{
Enforce that columns in any data sheet have definitions in Definitions sheet
}
