% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_enforce_mandatory_columns_present}
\alias{template_helper_enforce_mandatory_columns_present}
\title{Enforce that mandatory columns listed in Definitions sheet are present in data}
\usage{
template_helper_enforce_mandatory_columns_present(data_df, definitions)
}
\description{
Enforce that mandatory columns listed in Definitions sheet are present in data
}
