% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_extract_definitions}
\alias{template_helper_extract_definitions}
\title{extract definitions corresponding to a specific sheet}
\usage{
template_helper_extract_definitions(sheetName, def)
}
\description{
Find the definitions for a specific sheet (e.g. Subjects, Samples)
by consulting the relevant column (e.g. attribute_in_Subjects, attribute_in_Samples)
}
