% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_extract_pipeline_meta_info}
\alias{template_helper_extract_pipeline_meta_info}
\title{Extract pipeline meta information}
\usage{
template_helper_extract_pipeline_meta_info(pipelines_df, workbook)
}
\arguments{
\item{pipelines_df}{data-frame containing rows in Pipelines sheet corresponding to a 
`[project_id, study_id, study_version]` record}
}
\description{
Given rows from Pipelines sheet, extract the unique values in `pipeline_information`
column, and find the matching rows in `pipeline-choices` sheeet
}
