% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_extract_record_related_rows}
\alias{template_helper_extract_record_related_rows}
\title{extract information pertaining to a project-study record}
\usage{
template_helper_extract_record_related_rows(workbook, sheetName, record)
}
\description{
Helper function for template Excel sheet.
}
\details{
Given a project-study record [project_id, dataset_id, dataset_version]
find the information from a target sheet (e.g. Subjects, Samples, Pipelines) pertaining
to that record
}
