% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_formulate_file_path}
\alias{template_helper_formulate_file_path}
\title{Formulate a file path from Excel template}
\usage{
template_helper_formulate_file_path(pipeline_df, local_path = TRUE)
}
\description{
Formulate a file path from Excel template
}
