% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_formulate_local_file_path}
\alias{template_helper_formulate_local_file_path}
\title{Formulate local file path from Pipelines sheet columns}
\usage{
template_helper_formulate_local_file_path(df_pipelines)
}
\arguments{
\item{df1}{data.frame of Pipelines sheet}
}
\description{
Formulate local file path from Pipelines sheet columns
}
