% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_suffix_by_entity}
\alias{template_helper_suffix_by_entity}
\title{Convert entity to suffix}
\usage{
template_helper_suffix_by_entity(entity)
}
\description{
=== Entity-name        Suffix ===
RNAQUANTIFICATION      RNA
VARIANT                DNA
FUSION                 DNA
COPYNUMBER_MAT         DNA
PROTEOMICS             Protein
}
