% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\docType{data}
\name{template_linker}
\alias{template_linker}
\title{Storing columns that link different sheets}
\format{An object of class \code{list} of length 3.}
\usage{
template_linker
}
\description{
what columns link FeatureSets between featureset_choices sheet and Pipelines sheet
`api_choices_col` is same as `choices_col` if not specified
}
\keyword{datasets}
