% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests-register.R
\name{test_non_null_ontology_elements}
\alias{test_non_null_ontology_elements}
\title{Test that all ontology entries are non-null}
\usage{
test_non_null_ontology_elements(df, suffix = c("_", "_term"))
}
\arguments{
\item{df}{data frame on which this test will be run}

\item{suffix}{suffix for ontology columns. Default: "_" or "_term".}
}
\description{
Test that all ontology entries are non-null
}
