% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{user_confirms_action}
\alias{user_confirms_action}
\title{check before proceeding}
\usage{
user_confirms_action(action)
}
\arguments{
\item{action}{a string describing the action that will be taken if user confirms}
}
\description{
ask user for confirmation before proceeding with an action
}
