% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_load-api.R
\name{api_register_measurements}
\alias{api_register_measurements}
\title{Register measurements}
\usage{
api_register_measurements(biosample_names, bios_df_ref, msmtset_df_ref)
}
\arguments{
\item{biosample_names}{\code{sample_name} column from rows of Pipeline sheet corresponding to current pipeline}

\item{bios_df_ref}{dataframe containing API entry for biosample data at current \code{dataset_id} (retrieve from reference object)}

\item{msmtset_df_ref}{dataframe containing API entry for current measurementset (retrieve from reference object)}
}
\description{
Excel sheet equivalent of \link{\code{populate_measurements()}}
}
