% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{autoconvert_char}
\alias{autoconvert_char}
\title{Auto-promote characters to types}
\usage{
autoconvert_char(df1, convert_logicals = TRUE)
}
\arguments{
\item{convert_logicals}{if TRUE, accept automatically converted logicals (see note above)}
}
\description{
Custom function
- should not autopromote non-character columns for VARIANT data
- for columns that are interpreted as logical, should use strict checking for TRUE/True, or FALSE/False
('T' is a valid character in genomic data)
}
