% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dao_scidb.R
\name{dao_get_entity}
\alias{dao_get_entity}
\title{Faster version of `get_project`}
\usage{
dao_get_entity(entity, con)
}
\description{
Instead of running the query per namespace, merge the namespaces into one array at query-time
The code is based off the code at `join_info_ontology_and_unpivot`
}
