% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dao_scidb.R
\name{dao_search_rnaquantification}
\alias{dao_search_rnaquantification}
\title{Faster implementation of `search_rnaquantification` for UI development}
\usage{
dao_search_rnaquantification(rnaquantificationset, feature, biosample_ref,
  dataset_lookup_ref, con)
}
\arguments{
\item{biosample_ref}{Reference dataframe containing all biosamples available to the user (retrieved by `dao_get_biosample(con = con)`)}
}
\description{
Faster implementation of `search_rnaquantification` for UI development
}
