% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{get_experiment}
\alias{get_experiment}
\title{retrieve all the experiments available to logged in user}
\usage{
get_experiment(con = NULL)
}
\description{
joins Measurement, Dataset (for study category field), and ExperimentSet arrays
to return experiment information to user
}
\examples{
experiments = get_experiment()
cat("Categorization of experiments by major type\\n")
table(experiments$measurement_entity)
cat("Categorization of experiments by sub type\\n")
table(paste(experiments$measurement_entity, experiments$name, sep = ": "))
cat("Categorization of experiments by sub type\\n")
table(experiments$`study category`)

}
