% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_individuals}
\alias{get_individuals}
\title{Retrieve individuals}
\usage{
get_individuals(individual_id = NULL, dataset_version = NULL,
  all_versions = FALSE, mandatory_fields_only = FALSE, con = NULL)
}
\description{
get_ENTITY can be used to retrive
\itemize{
\item one individual (e.g. \code{individual_id = 33})
\item more than one individual (e.g. \code{individual_id = c(33, 44)})
\item all individuals visible to user (e.g. \code{get_individuals()})
}
}
