% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_db}
\alias{init_db}
\title{Function to (potentially delete and) initialize arrays in GA4GH respository}
\usage{
init_db(arrays_to_init)
}
\description{
The function requires that you are connected to SciDB.
}
\examples{
\dontrun{
init_db('FUSION')
init_db(c('FUSIONSET', 'FUSION'))
init_db(arrays = get_entity_names()) # Warning! This reinitializes all the arrays
}
}
