% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_db}
\alias{init_db}
\title{(potentially delete and) initialize arrays managed by Insight API}
\usage{
init_db(arrays_to_init, con = NULL)
}
\description{
The function requires that you are connected to SciDB as an user with admin privileges
}
\examples{
\dontrun{
init_db('FUSION')
init_db(c('MEASUREMENTSET', 'FUSION'))
init_db(arrays = get_entity_names()) # Warning! This reinitializes all the arrays
}
}
