% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_info_ontology_and_unpivot}
\alias{join_info_ontology_and_unpivot}
\title{Join flex fields, unpivot and convert ontology fields}
\usage{
join_info_ontology_and_unpivot(qq, arrayname, mandatory_fields_only = FALSE,
  replicate_query_on_info_array = FALSE, con = NULL)
}
\arguments{
\item{replicate_query_on_info_array}{see description at [join_info()]}
}
\description{
Join flex fields, unpivot and convert ontology fields
}
