% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{myExcelLoader}
\alias{myExcelLoader}
\title{Helper function to load Excel workbook into memory}
\usage{
myExcelLoader(filename)
}
\arguments{
\item{filename}{path to Excel workbook}
}
\description{
Helper function to load Excel workbook into memory
}
