% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{myExcelLoader}
\alias{myExcelLoader}
\title{Load Excel workbook into memory}
\usage{
myExcelLoader(filename, use_readxl = TRUE)
}
\arguments{
\item{filename}{path to Excel workbook}

\item{use_readxl}{use readxl package (default), otherwise XLConnect package}
}
\description{
To be used in conjunction with `myExcelReader()`
`
}
