% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{myExcelReader}
\alias{myExcelReader}
\title{Read sheet from Excel workbook}
\usage{
myExcelReader(workbook, sheet_name)
}
\arguments{
\item{workbook}{Workbook must be of type \code{XLConnect::workbook} (loaded by \code{XLConnect::loadWorkbook()})
or of type \code{list} (loaded by \code{revealgenomics::myExcelLoader()})}

\item{sheet_name}{name of sheet to read from workbook}
}
\description{
To be used in conjunction with \code{myExcelLoader()}
}
\details{
Note usage of the useCachedValues parameter
otherwise, was throwing warnings about the concatenation field
}
