% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{myExcelReader}
\alias{myExcelReader}
\title{Helper function to read sheet from Excel workbook}
\usage{
myExcelReader(workbook, sheet_name)
}
\description{
note usage of the useCachedValues parameter
otherwise, was throwing warnings about the concatenation field
}
