% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{populate_measurements}
\alias{populate_measurements}
\title{register all measurements for a specific dataset}
\usage{
populate_measurements(dataset_id, dataset_version, con = NULL)
}
\description{
The function should be called after loading the \code{MeasurementData} arrays\code{(e.g. Variant, RNA-seq etc.). The}MEASUREMENT\code{entity records one entry for each entry in a}MeasurementData` array per biosample,
per pipeline
}
\details{
(A \code{MEASUREMENT} combines both Experiment and Pipeline information)
}
