% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{pretty_print}
\alias{pretty_print}
\title{Pretty print a large vector of strings, integers etc.}
\usage{
pretty_print(vec, prettify_after = 7)
}
\arguments{
\item{vec}{vector that is to be pretty printed}

\item{prettify_after}{prettify output if length of vector is longer than this limit}
}
\description{
Pretty print a large vector of strings, integers etc.
}
