% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_excel_loader.R
\name{register_entities_excel}
\alias{register_entities_excel}
\title{register entities via Excel sheet}
\usage{
register_entities_excel(study_worksheet, ...)
}
\arguments{
\item{study_worksheet}{path to Excel worksheet}

\item{...}{refer parameters for function \code{\link{register_entities_workbook}}}
}
\description{
Register entities using Excel sheet.
Excel sheet has complete data about metadata (\code{PROJECT}, \code{DATASET}, \code{INDIVIDUAL}, ...) ,
and links to files storing the \code{MeasurementData} entities
(e.g. \code{RNASeq}, \code{Variant} etc.)
}
