% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_excel_loader.R
\name{register_entities_excel}
\alias{register_entities_excel}
\title{register entities via Excel sheet}
\usage{
register_entities_excel(study_worksheet, register_upto_entity = c("all",
  .ghEnv$meta$arrOntology, .ghEnv$meta$arrDataset, .ghEnv$meta$arrDefinition),
  BASEPATH = "", con = NULL)
}
\arguments{
\item{register_upto_entity}{use this function to load up till user-specified entity.
e.g. you can load `Ontology`, `(Project)` and `Dataset`, 
`Definition` and not load the rest. 
Default value is `all` i.e. all entities are loaded}
}
\description{
Register entities using Excel sheet. 
Excel sheet has complete data about metadata (`PROJECT`, `DATASET`, `INDIVIDUAL`, ...) ,
and links to files storing the `MeasurementData` entities 
(e.g. `RNASeq`, `Variant` etc.)
}
