% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-measurement-upload.R
\name{register_expression_dataframe}
\alias{register_expression_dataframe}
\title{Upload data into gene expression, protein expression arrays}
\usage{
register_expression_dataframe(df1, dataset_version, con = NULL)
}
\description{
Upload data into gene expression, protein expression arrays
}
