% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_expression_by_one_or_more_biosamples}
\alias{search_expression_by_one_or_more_biosamples}
\title{Subpath for \code{search_expression}}
\usage{
search_expression_by_one_or_more_biosamples(entity, biosample,
  formExpressionSet = FALSE, con = NULL)
}
\description{
implementation of \code{search_expression} when searching by one or more biosamples. Here you are not selecting by MeasurementSet-s or Features.
}
