% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_fusion}
\alias{search_fusion}
\title{Search Fusion data}
\usage{
search_fusion(measurementset, biosample = NULL, feature = NULL,
  dataset_lookup_ref = NULL, con = NULL)
}
\arguments{
\item{inferSchemaOnQuery}{optimization flag (default: TRUE. if FALSE, uses schema hand-baked in internal code. 
Use with caution until schema is automatically inferred from YAML file)}
}
\description{
Search Fusion data
}
