% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_measurementdata}
\alias{search_measurementdata}
\title{Unified function to download entire pipeline worth of data}
\usage{
search_measurementdata(measurementset, con = NULL)
}
\arguments{
\item{measurementset}{dataframe containing one row of pipeline information}

\item{con}{connection object (optional  if using \code{rg_connect()})}
}
\description{
Unified function to download entire pipeline worth of data
}
