% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_variants}
\alias{search_variants}
\title{search variants}
\usage{
search_variants(measurementset, biosample = NULL, feature = NULL,
  autoconvert_characters = TRUE, variants_in_one_array = TRUE, con = NULL)
}
\arguments{
\item{measurementset}{pipeline dataframe (mandatory)}
}
\description{
Search variants by pipeline, sample and/or  feature
}
