% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2_template-helper.R
\name{template_helper_extract_pipeline_meta_info}
\alias{template_helper_extract_pipeline_meta_info}
\title{Extract pipeline meta information}
\usage{
template_helper_extract_pipeline_meta_info(pipelines_df, choicesObj)
}
\arguments{
\item{pipelines_df}{data-frame containing rows in Pipelines sheet corresponding to a 
`[project_id, study_id, study_version]` record}

\item{choicesObj}{list containing instantiated objects of PipelineChoices, FilterChoices
and FeaturesetChoices classes}
}
\description{
Given rows from Pipelines sheet, extract the unique keys for pipeline, filter and featurset-s.
Then find the matching and relevant rows in `pipeline_choices`, `filter_choices` and 
`featureset_choices`
}
